Add flash messages for order. Update like button
Order button now has a flash message indicating if order was
created, and a warning if fields are incomplete.
Like button is just a regular button with a like text on it.
For some crazy reason, deployed site is not rendering images from
local images even though it does in localhost! So got rid of that!!